+++ internal/cmd/internal/olmv1/operator_uninstall.go
@@ -4,12 +4,12 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/operator-framework/kubectl-operator/internal/cmd/internal/log"
-	experimentalaction "github.com/operator-framework/kubectl-operator/internal/pkg/experimental/action"
+	v1action "github.com/operator-framework/kubectl-operator/internal/pkg/v1/action"
 	"github.com/operator-framework/kubectl-operator/pkg/action"
 )
 
 func NewOperatorUninstallCmd(cfg *action.Configuration) *cobra.Command {
-	u := experimentalaction.NewOperatorUninstall(cfg)
+	u := v1action.NewOperatorUninstall(cfg)
 	u.Logf = log.Printf
 
 	cmd := &cobra.Command{
