+++ pkg/action/config.go
@@ -11,7 +11,7 @@ import (
 
 	v1 "github.com/operator-framework/api/pkg/operators/v1"
 	"github.com/operator-framework/api/pkg/operators/v1alpha1"
-	olmv1 "github.com/operator-framework/operator-controller/api/v1alpha1"
+	olmv1 "github.com/operator-framework/operator-controller/api/v1"
 	operatorsv1 "github.com/operator-framework/operator-lifecycle-manager/pkg/package-server/apis/operators/v1"
 )
